refactor: MoneyReceiver 생성
