+++ src/main/java/lotto/domain/WinningNumbers.java
@@ -1,0 +1,2 @@
+package lotto.domain;public class WinningNumbers {
+}
