+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -1,0 +1,2 @@
+package lotto.domain;public class WinningNumbersTest {
+}
